+++ test/run_triggers_test.py
@@ -42,7 +42,8 @@ class TestTFCRunTriggers(TestTFCBaseTestCase):
         ]
 
         # List the triggers and confirm there are none
-        some_triggers = self._api.run_triggers.list(self._target_ws_id, filters=test_filters)["data"]
+        some_triggers = self._api.run_triggers.list(\
+            self._target_ws_id, filters=test_filters)["data"]
         self.assertEqual(len(some_triggers), 0)
 
         # Create a trigger
@@ -58,11 +59,13 @@ class TestTFCRunTriggers(TestTFCBaseTestCase):
                 }
             }
         }
-        created_trigger = self._api.run_triggers.create(self._target_ws_id, create_payload)["data"]
+        created_trigger = self._api.run_triggers.create(\
+            self._target_ws_id, create_payload)["data"]
         created_trigger_id = created_trigger["id"]
 
         # List the triggers again and confirm there is one
-        some_triggers = self._api.run_triggers.list(self._target_ws_id, filters=test_filters)["data"]
+        some_triggers = self._api.run_triggers.list(\
+            self._target_ws_id, filters=test_filters)["data"]
         self.assertEqual(len(some_triggers), 1)
 
         # List all triggers and confirm there is one
@@ -76,5 +79,6 @@ class TestTFCRunTriggers(TestTFCBaseTestCase):
 
         # Destroy the run trigger, confirm that we have zero run triggers again
         self._api.run_triggers.destroy(created_trigger_id)
-        some_triggers = self._api.run_triggers.list(self._target_ws_id, filters=test_filters)["data"]
+        some_triggers = self._api.run_triggers.list(
+            self._target_ws_id, filters=test_filters)["data"]
         self.assertEqual(len(some_triggers), 0)